Run the Sequencer and Synth from the audio thread
SwapBuffers blocks the UI thread unpredictably and we can't fill the
buffer from there. So start moving this logic to the audio thread.
The script is a bit heavy and could cause underruns by waiting for that
callback, but worst case we can move this to a third thread later on if
scripts start becoming heavier.