Remix gb channels with alignments for the scope visualization
The audio mixing is a stream of unaligned waveforms and this means that
when multiple channels are played together some of them will not have
their shape clearly visible in the scope.

To make this a bit easier on the eyes, re-mix all channels for visualiztion
purposes after aligning them by the start of their phase.