Disable the visualization scroll animation
This broke with slint 0.2.5 because of a timing fix and
it was almost by luck that it initially worked.
The whole animation 1/4s in advance thing was weird anyway
and it would be more reliable to keep the panel animating
but instead get the scroll position from the UI to decide
where to put it on the timeline.