Also auto-scroll the song patterns and instruments according to the selection on desktop
Using a Flickable both to scroll with the mouse and from the selection doesn't
play well with bindings. I'd need some sort of dynamic bindings that are
re-created after the user somehow re-snap the viewport to the selection.

For now just scrolling causing the selection to change seems to at least
allow working with the thing.