Fix the instruments grid disapearing on window blur
has_focus is also false when the window doesn't have focus,
this means that we can't use the FocusScope as state for whether
the instruments grid is visible.

Go back to a separate enum for the focused panel, this also makes
it impossible to change the focus with the mouse, but that was actually
annoying as it would require to click twice.