Reduce the fixed point precision to 8 bits
The maximum supported frequency of 131072 by the GBA takes 17 bits, this
means that with an u32 we have 15 bits left before we overflow.
To allow multiplying by an int factor and then divide to simulate floating
point multiplications, using 10 bits would mean that we only have 5 bits
left up.

So reduce the precision to 8 bits so that we have 7 bits up for the highest
frequencies (and a bit more for more common lower octaves).
8 bits is still enough to represent NOTE_FREQUENCIES without loss vs a
floating point when ultimately encoded to GBA frequencies.