gba: Implement sync mode using direct sound
This uses the direct sound channel to generate the sync pulse using DMA
at low frequency to avoid using too many CPU cycles.

This however doesn't work out of the box on the GBA and 3DS as the
volume/voltage is too low on the left channel to trigger a tic on my PO
devices.

It's however possible to raise the volume above the physical slider's
level on the 3DS through the open_agb_firm configuration. With the
maximum volume set there it seems to work well.