gba: Use both L and R buttons as Shift
Now that the menu is implemented on the GBA with the save function,
the L button can also be used as Shift to make it possible to use
either hand since it can be combined with buttons from both sides.

The different layout of select/start buttons on different GBA models
also make it difficult to have this button only on one side.