Display instrument parameters as unsigned hexadecimal on desktop as on the GBA
This was used on the GBA to avoid having to reserve two columns for the sign character,
but having this displayed differently on desktop makes little sense.
It's also nice to have the option of using each i8 parameter to represent two u4 values.

It's quite crappy that the negative values are shown in the UI as 1-complement unsigned,
but I think it's still useful to have the left-right cycle of param values to go down
to -128 instead of blocking at 0, so let's try it this way and see whether it is still
better than switching to a u8 also in the cycle implementation and in the instruments
interface.